requirements.txt
Django==4.1
django-nested-admin==3.4.0
fpdf==1.7.2
django-nested-formset==0.1.4
requests==2.28.1
python-magic
Pillow
django-bootstrap-v5
